Monkey patch rails to make system table query efficient
Rails tries to query the system tables to find the primary key
and it's controlling sequence, but it doesn't specify that it is
looking for objects of class "pg_class" so the pg_depend table
has to be sequentially scanned instead of being accessed using
the index.

Tests on the production database show that the time taken for
this query drops from 3.5s to 15ms if the index is used.