Add _html suffixes to some resource strings which have changed
Where a translatable string contains (directly or indirectly via
interpolations) some HTML it is better to add an _html suffix to
the key rather than wrap the result in raw so that escaping is
handled automatically for interpolations.

It's not worth doing until a string is changed and needs to be
retranslated though, as it will just break translations.