Single tab keypress focus on map for keyboard users
When one goes to the default map page, the first tabindex is set to the
search field, which is good. But, to use the keyboard with the map (arrows
for scrolling, =/- for zooming), one must press the tab key many times.

This modification quickly gets the keyboard to a much-used location via
a single keypress.