Do not fit map to search result (object) twice
When clicking on search results the respective osm object is loaded
asynchronously and afterwards shown on the map. Some objects can take
a few seconds to load however.

If a user already panned/zoomed the map manually during this time then
their actions are reverted because the map fits to the object again.

There is actually no need for this behaviour, because a good geocoder
already returns an appropriate bounding box for the search result.