Make API and web roll back any open transactions on timeout.
This seems like it would be the default behaviour, but because the exception that Timeout::timeout raises on the other thread doesn't appear to be rescue-able, Rails doesn't detect it. However, it still runs the ensure block, which attempts to commit the transaction.

The result of this is that a long-running transaction can potentially be interrupted while partially complete and will commit the partial result rather than rolling back.