Allow searching via UK GSS code.
The UK Office for National Statistics maintains a series of codes to
represent geographical areas of the UK. These codes follow a standard
format and are recorded in OpenStreetMap using the ref:gss tag.

This change lets map users search the map by UK GSS code. It uses the
Overpass API to look for a relation with the code provided by the user.

Example search: E06000001 will return Hartlepool (unitary authority).