Replace tests with recreated subset of tests
These new, simple tests focus mainly on demonstrating that the redaction
is not allowed to be saved to the database when the title
and description are blank or made up of spaces. Eliminated comments that
didn't seem to be relevant any longer and rewrote the tests to avoid
any usage of session or controller. Also, embraced the create function
to take on a more factory like approach of creating redactions for testing.