Make 'back' button not look disabled.
When viewing a message in the inbox, the 'back' button looks paler
than the other buttons below the message. This can be mistaken
for that button to be disabled, and thus there being no way
(other than the browser's own 'back' button or accessing it
anew from the logged-in-user drop down menu) to get back
to the list of messages in the inbox. (Actually, the button
is just 'deemphasized', whatever that's supposed to mean.)

This change makes it no longer deemphazized, thus displaying
it in the same style as the other buttons there.