Fix installation instructions for Mac OS X
The installation of `libxml-ruby` was broken because it was picking up the system version of the libxml2 library. By setting a `bundler` config item, it's possible to point it to the right place.

Also updated the instructions for `phantomjs`, which Homebrew has moved to a cask (and is now not officially supported), and added some of the extra command line tools that are used (`pngcrush`, `optipng`, etc...)