Rename nav-related body classes
This makes it clearer that the classes are only for navigation menu
matters. Other things should be done on media queries, or ideally,
using bootstrap components like grids.

This has a side effect of removing the smaller font size from the
body when the navigation menu is in the small-nav state.