Style changeset descriptions with italics on browse pages
Alternative to #580

This reworks changeset descriptions to use italics instead of header
elements, in order to make them more like quotes.

Care has been taken to make sure that there is still a header element
as the first element of the common details partial, in order to make
the history pages look right (where the partial appears multiple times).