WIP: Sync more hash-url-params for iD
This commit includes a few console.logs since it is still WIP.

It solved part of the issue described in https://github.com/openstreetmap/openstreetmap-website/issues/3031. Whenever I pan the map in iD, the osm-hash-params get updated witht the current iD-in-iframe-configuration. However, click actions in the iD iframe dont trigger this update, only panning the map.