Remove conditional for polyline layer
When a user searches for a path and one is not found the polyline layer is
removed. When a user changes the mode of transportation the event listener will
not reroute if there is no polyline layer. This commit removes the conditional
that prevents the routing code from running after the polyline has been
removed