Bootstrap 5: Remove the padding from trace nav buttons on larger sizes
If additional py is added when the tabs are laid out in full, then the
other tabs start to become detatched from the rest of the page and a
line appears under the active tab.

We don't need the padding in the horizontal layout, so remove it, and
tweak the size of the padding when collapsed.