Combine small.scss into common.scss
We've moved away from the idea of having extra rules for small screens,
and instead follow the bootstrap convention of designing for small
screens first and overriding those rules if needed for larger screens.

Having a separate file remaining for a limited number of small screen
rules is therefore counter-productive. Moving the remaining rules into
common.scss, adjacent to the related rules for larger screen, makes
the behaviour easier to understand, as well as making it easier to
refactor the rules to a small-screen-first approach.