Organizers can edit their communities.  Admin can promote members to organizers (first admin).
Allow non-admins to create a community and become the organizer of it.

Rename community.organizer to community.leader.  There is just one leader.  The leader also is a community_member with the role of organizer.  This should have been a separate commit.  Sorry.

Change access control so any organizer can modify communities, links, and members.

Allow organizers to change the role of a member.

I pulled in lots of later patches to get 100% coverage of passing tests.

0014, 0046, 0118, 0158, 0204, 0206

Use "could not be" phrasing.

0178

0238

Not all members are organizers

0023

0172

0166

0167

Write up a test for update microcosm as a non-organizer.

Factor out ability condition.

0215

0053

0222