Delete config/locales/zh-HK.yml
We have `zh-HK.yml`, which I believe should be the correct name for a locale file, but it is not receiving updates. We have `zh-hk.yml` which is receiving updates. We should not have both in a world where some OSs are case sensitive and some are.

HK should be deleted and whatever upstream is using hk should be corrected.