Improve scaling and vertical alignment of browse icons
Although we're using SVG icons because they are being added using
pseudo elements it's currently impossible to directly set a size for
them to scale to as https://stackoverflow.com/a/8993934 explains.

So instead we use a solution based on the alternate approach there
of setting the image as a background image, which can be scaled to
our preferred size and then aligned.