iD: listen to  manual hashchanges from iframe parent
Pan to the new `map` location when the hash change was not not triggered by map interaction from inside iD itself.

This mirrors the behaviour when manually editing the `map` hash parameter on osm.org outside of iD.

This also fixes https://github.com/openstreetmap/iD/issues/10592 (error in js console when iD is opened outside of an iframe by directly navigating to osm.org/id).