Hopefully the final commit for these files. My choices try to address two pressures.
It is easier to track upstream changes to openstreetmap-website when there's less potential for
merge conflicts. settings.yml is now identical to upstream and all OHM modifications are
made in settings.local.yml; that file is not checked in to source control but a fleshed-out
skeleton is available at example.settings.local.yml. What information is in there is publicly
available with the exception of OAuth IDs, and those are generated per installation.

These two files should be identical to those in images/web/config/ of the ohm-deploy repo and
they are manipulated by the start.sh script in the parent directory. Those scripts have been
modified so that they use the proper settings file for variable substitution.