Restrict version of dry-schema for ruby 3.1.2 compatibility
https://bugs.ruby-lang.org/issues/18673 affects some methods using
anonymous blocks, and was fixed in ruby 3.1.3. However, debian/bookworm
ruby 3.1 package is based on 3.1.2 and does not yet contain this fix.

Since bookworm is used in our Docker build, and also used in production
by OSMF, then we can restrict dry-schema to a version that does not
use anonymous blocks.

Fixes #5482