Remove html-safe/unsafe branches from rich text linkify
Rinku outputs html. It would make no sense to escape Rinku's output, the output should be safe when it's created. To make sure that it's safe, we need to escape the input. As it is currently used in osm-website, the input is already html-safe and escaping does nothing. However it's safer to escape anyway in case the future usage changes.