Reset sidebar scroll position on map pan in place history
Otherwise the user may not notice that the list was completely reloaded. Also it makes sense to start looking at the top of the reloaded list. Without resetting the scroll position, the bottom is likely to be shown if the user went on to browse the list and pressed "Load more" a few times.