Don't do partial renders for turbo page visits
Setting turbo-action=advance as our pagination helper does causes
a frame navigation to be promoted to a page visit which means turbo
will try and merge the head elements and remove meta and link elements
which are missing from the turbo response.

In order to avoid breaking things we need to ensure we send the full
head element in such cases or things like the CSRF token will be lost.

Fixes #6172