Add and unignore .ruby-version file to ensure consistent dev Ruby
Having a .ruby-version file has become a somewhat standard practice in Ruby development such that the chruby CLI tool and the ruby-lsp VSCode extension expect it to exist.

While not officially signed off on, there's even been a proposal for a spec for the .ruby-version file format over in rubygems: https://github.com/orgs/rubygems/discussions/7074.