Enable ActionMailer Previews
The previews run using the development database, which may or may not already contain
sufficient records (e.g. diary comments). So to avoid failures, we generate content
using factory bot, and to avoid filling your development database with duplicate data,
we run the previews inside a transaction.

If you manually create records in the database that collide with factory bot (e.g. users
with the same email pattern as the factory) then you'll need to update your manual records
to avoid collisions.